Added methods for redirect on 404 errors. Also added navbar.

+++ app/Http/Controllers/PostsController.php
@@ -6,35 +6,23 @@ use Illuminate\Http\Request;
 
 use App\Http\Requests;
 use App\Http\Controllers\Controller;
-use App\Models;
 use App\Models\Post;
+use Session;
+use Log;
 
 class PostsController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
+
 
     public function index()
     {
-        // show all the posts
-        $posts = Models\Post::paginate(4);
+        $posts = Post::paginate(4);
         return view('posts.index')->with('posts', $posts);
-
-
     }
 
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
+
     public function create()
     {
-        // show the form to create a post
-
         $data = [];
         $posts['title'] = '';
         $posts['content'] = '';
@@ -44,88 +32,75 @@ class PostsController extends Controller
         return view('posts.create', $data);
     }
 
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
+
     public function store(Request $request)
     {
-        // save the new post
         $request->session()->flush();
 
         $this->validate($request, Post::$rules);
 
-        $post = new Models\Post();
+        $post = new Post();
 
         $post->title = $request->title;
         $post->content = $request->content;
         $post->url = $request->url;
         $post->created_by = 1;
+        Log::info("Post Created \n \tTitle: {$post->title} \n \tContent: {$post->content} \n \tURL: {$post->url} \n \tCreated By: {$post->created_by}");
         $post->save();
 
-        \Session::flash('successMessage', 'Post saved successfully!');
+        Session::flash('successMessage', 'Post saved successfully!');
 
 
         return redirect()->action('PostsController@show', $post->id);
     }
 
-    /**
-     * Display the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
+
     public function show($id)
     {
-        // show a specific post based on the id
-        $posts = Models\Post::find($id);
+        $posts = Post::find($id);
 
 
         if (!$posts) {
-            \Session::flash('errorMessage', 'Post not found.');
-            return redirect()->action('PostsController@index');
+            Log::info("Post with id of $id cannot be found.");
+            abort(404);
         }
         
         return view('posts.show')->with('posts', $posts);
     }
 
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
+
     public function edit($id)
     {
-        // Show the form for editing the specified resource.
-        $posts = Models\Post::find($id);
+        $posts = Post::find($id);
+
+        if (!$posts) {
+            Log::info("Post with id of $id cannot be found.");
+            abort(404);
+        }
 
         return view('posts.edit')->with('posts', $posts);
     }
 
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
+
     public function update(Request $request, $id)
     {
 
         $this->validate($request, Post::$rules);
 
-        // Update the post in the database
-        $post = \App\Models\Post::find($id);
+
+        $post = Post::find($id);
+
+        if (!$post) {
+            abort(404);
+        }
+
         $post->title = $request->title;
         $post->content = $request->content;
         $post->url = $request->url;
         $post->save();
 
         if ($post) {
-            \Session::flash('updateMessage', 'Post successfully updated.');
+            Session::flash('updateMessage', 'Post successfully updated.');
             return redirect()->action('PostsController@show', $post->id);
         }
 
@@ -133,19 +108,21 @@ class PostsController extends Controller
         return PostsController::show($id);
     }
 
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
+
     public function destroy($id)
     {
-        // delete a post based on the id
-        $post = Models\Post::find($id);
+        $post = Post::find($id);
+
+        if (!$post) {
+            Log::info("Post with id of $id cannot be found.");
+            abort(404);
+        }
+
+        Log::info("Post Deleted \n \tTitle: {$post->title} \n \tContent: {$post->content} \n \tURL: {$post->url} \n \tCreated By: {$post->created_by}");
+
         $post->delete();
 
-        \Session::flash('deleteMessage', 'Post successfully deleted.');
+        Session::flash('deleteMessage', 'Post successfully deleted.');
 
         return redirect()->action('PostsController@index');
     }
